sample_code/python/tcp_v2/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
backoff>=1.10.0
loguru>=0.5.3
pyais
PyYAML>=5.4.1
google-cloud-bigquery
google-api-core